project - timemarker

+++ src/models/ProjectProject.js
@@ -64,7 +64,7 @@ const ProjectProject = sequelize.define('ProjectProject', {
     allowNull: false
   },
   burn_minutes: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.DECIMAL(15, 6),
     allowNull: false,
     defaultValue: 0
   },
